moved fastapi creation and endpoints to api.py
